warning clean up

+++ triador.cpp
@@ -154,7 +154,7 @@ void execute(const vector<string> &opcodes, const vector<int> &opargs) {
         PC++;
         if (string("JP") == oc) {
             PC = (R[12]*27 + arg)+364;
-            assert(PC<opcodes.size());
+            assert((size_t)PC<opcodes.size());
         }
         display_memory_state();
     }
